refactor(cubeapi): extract ExampleService and modularise example registry
