+++ CubeAPI/src/services/mod.rs
@@ -4,6 +4,7 @@
 
 pub mod cluster;
 pub mod exec;
+pub mod examples;
 pub mod sandboxes;
 pub mod snapshots;
 pub mod templates;
@@ -84,6 +85,7 @@ fn is_valid_dns_domain_name(domain: &str) -> bool {
 #[derive(Clone)]
 pub struct AppServices {
     pub cluster: cluster::ClusterService,
+    pub examples: examples::ExampleService,
     pub exec: exec::ExecService,
     pub sandboxes: sandboxes::SandboxService,
     pub snapshots: snapshots::SnapshotService,
@@ -98,6 +100,15 @@ impl AppServices {
     ) -> Self {
         Self {
             cluster: cluster::ClusterService::new(cubemaster.clone()),
+            examples: examples::ExampleService::new(
+                config.cube_api_url.clone(),
+                config.default_template_id.clone(),
+                config.cube_proxy_node_ip.clone(),
+                config.cube_proxy_port_http,
+                config.sandbox_domain.clone(),
+                config.sandbox_proxy_url.clone(),
+                config.envd_auth.clone(),
+            ),
             exec: exec::ExecService::new(
                 http_client,
                 config.sandbox_proxy_url.clone(),
